[FIX] base.trplot(): argument length was scaling hom. tf. fixes #2
HINT: The issue arises whenever the passed pose T in SE3 is not in the origin @ [0;0;0] and the length != 1.
The reason is that the length is not just defining the length of the unit-vector in the x,y,z-direction,
but also the scaling the whole Pose (it is rather a SIM3 Matrix than a homogeneous matrix).